requirements.txt
# Dependencias del proyecto Chess Trainer

# Framework Web
fastapi>=0.115.0
uvicorn>=0.32.0

# HTTP Client
httpx>=0.27.0

# Configuración
PyYAML>=6.0.2
python-dotenv>=1.0.0
Jinja2>=3.1.0

# JSON Path para parsing de APIs REST
jsonpath>=0.82.2

# Validación y manipulación de ajedrez
python-chess>=1.10.0

# Logging mejorado (opcional pero recomendado)
colorlog>=6.8.0

# Para motores generativos (LLM)
# Nota: Descomentar según el proveedor que uses

# OpenAI (para GPT-4, GPT-3.5, etc.)
# openai>=1.30.0

# Anthropic (para Claude)
# anthropic>=0.25.0

# Cohere (para modelos Cohere)
# cohere>=5.0.0

# LangChain (para agentes y pipelines LLM)
# langchain>=0.2.0
# langchain-openai>=0.1.0
# langchain-anthropic>=0.1.0

# LangGraph (para agentes con estado y grafos)
# langgraph>=0.1.0

# Haystack (framework alternativo para LLM)
# haystack-ai>=2.0.0

# Transformers (para modelos locales)
# transformers>=4.40.0
# torch>=2.3.0

# Ollama (para modelos locales con Ollama)
# ollama>=0.2.0

# Para desarrollo y testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
